refactor: StringParser 클래스의 패키지를 utils 로 변경
